requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
base58==2.1.1
bitarray==1.2.2
certifi==2021.10.8
charset-normalizer==2.0.11
cytoolz==0.11.2
eip712-structs==1.1.0
eth-abi==2.1.1
eth-account==0.5.7
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
frozenlist==1.3.0
hexbytes==0.2.2
idna==3.3
iniconfig==1.1.1
ipfshttpclient==0.8.0a2
jsonschema==3.2.0
lru-dict==1.1.7
multiaddr==0.0.9
multidict==6.0.2
netaddr==0.8.0
packaging==21.3
parsimonious==0.8.1
pluggy==1.0.0
protobuf==3.19.4
py==1.11.0
py-order-utils>=0.3.0
pycryptodome==3.14.0
pyparsing==3.0.7
pyrsistent==0.18.1
pysha3==1.0.2
pytest==6.2.5
python-dotenv==0.19.2
requests==2.27.1
rlp==2.0.1
six==1.16.0
toml==0.10.2
toolz==0.11.2
urllib3==1.26.8
varint==1.0.2
web3==5.27.0
websockets==9.1
yarl==1.7.2
black==22.3.0
flask
